requirements.txt
# --- LLM Integration ---
openai>=1.82.0
google-generativeai>=0.3.0

# --- GUI Automation ---
pyautogui>=0.9.54
pygetwindow>=0.0.9
psutil>=5.9.0
pynput>=1.7.6
pyautogui
pywin32

# --- Computer Vision ---
opencv-python>=4.8.0.76
ultralytics>=8.0.20  # YOLOv8
transformers>=4.39.3  # For DETR
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0

# --- UI ---
streamlit>=1.30.0

# --- Config, Parsing & Logging ---
pydantic>=2.0.0
PyYAML>=6.0.2
loguru>=0.7.2

# --- Utilities ---
requests>=2.31.0

# --- Testing ---
pytest>=7.4.0
pytest-mock>=3.11.1

omniparse>=1.0.0

